+++ src/main/resources/application-local.yml
@@ -7,4 +7,4 @@ spring:
 
 mercadopago:
   access_token: ${MP_ACCESS_TOKEN:TEST-2194493215547460-072323-e7020fc0d11389fdae3dc86c6b9a1e4f-115466963}
-  notification_url: ${MP_NOTIFICATION_URL:https://webhook.site/820da4fd-08cc-446d-8f0d-d4b9f22738b9/pagamentos/pedidos/notifications}
+  notification_url: ${MP_NOTIFICATION_URL:https://614f-2804-7f0-bec0-2fa7-4d18-dc53-6657-d0ff.ngrok-free.app/pagamentos/pedidos/notifications}
