+++ src/main/resources/application.yml
@@ -32,4 +32,4 @@ springdoc:
 
 mercadopago:
   access_token: ${MP_ACCESS_TOKEN:TEST-2194493215547460-072323-e7020fc0d11389fdae3dc86c6b9a1e4f-115466963}
-  notification_url: ${MP_NOTIFICATION_URL:https://webhook.site/3defc4e7-bb96-4910-8d7a-cb50eb7c5dcf/pedidos/mercadopago/notifications}
+  notification_url: ${MP_NOTIFICATION_URL:https://614f-2804-7f0-bec0-2fa7-4d18-dc53-6657-d0ff.ngrok-free.app/pedidos/mercadopago/notifications}
